+++ src/types/mod.rs
@@ -1,7 +1,9 @@
+mod bitrate;
 mod extension;
 mod metadata;
 mod timestamp;
 
+pub use bitrate::Bitrate;
 pub use extension::Extension;
 pub use metadata::Metadata;
 pub use timestamp::{Timestamp, Timestamps};